server-py/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
firebase-admin==6.5.0
PyJWT==2.9.0
